refactor(components.fileeditor): move to webUniverseComponents
ref: OM-157
