+++ client/app/core/translations/Messages_fr_FR.xml
@@ -247,14 +247,11 @@
    <translation id="common_other" qtlid="19963">Autre</translation>
    <translation id="common_yes" qtlid="34776">Oui</translation>
    <translation id="common_no" qtlid="34788">Non</translation>
-   <translation id="common_add" qtlid="37923">Ajouter</translation>
    <translation id="common_delete" qtlid="2414">Supprimer</translation>
    <translation id="common_configure" qtlid="4010">Configurer</translation>
    <translation id="common_delete_in_progress" qtlid="200300">Suppression en cours...</translation>
-   <translation id="common_delete_all" qtlid="159891">Tout supprimer</translation>
    <translation id="required" qtlid="182514">Champs requis</translation>
    <translation id="required_fields" qtlid="433550">Les champs suivis d'un astérisque sont obligatoires.</translation>
-   <translation id="common_go_to_next_error" qtlid="282562">Aller à l'erreur suivante</translation>
    <translation id="common_skip_to_main_content" qtlid="454052">Aller au contenu principal</translation>
 
    <translation id="common_menu_support_assistance" qtlid="52699">Assistance</translation>
