+++ client/app/css/source.less
@@ -59,8 +59,7 @@
 @import (less) "@ovh-ux/ovh-utils-angular/bin/template/wizard/wizardStep/wizardStep.css";
 @import (less) "font-awesome/css/font-awesome.min.css";
 @import (less) "animate.css/animate.min.css";
-@import (less) "@bower_components/angular-xeditable/dist/css/xeditable.min.css";
-@import (less) "../components/fileEditor/file-editor.css";
+@import (less) "angular-xeditable/dist/css/xeditable.min.css";
 @import (less) "ovh-angular-sso-auth-modal-plugin/dist/ovh-angular-sso-auth-modal-plugin.min.css";
 @import (less) "ovh-angular-responsive-popover/dist/ovh-angular-responsive-popover.min.css";
 @import (less) "ovh-angular-tail-logs/dist/ovh-angular-tail-logs.css";
