+++ client/app/email-domain/mailing-list/moderators/delete/email-domain-mailing-list-moderators-delete.html
@@ -19,12 +19,12 @@
                     <div class="form-group" data-ng-class="{ 'has-error': ctrl.errors.length > 0 }">
                         <label class="control-label required"
                                data-translate="mailing_list_tab_delete_moderators_list"></label>
-                        <div data-file-editor
-                             data-file-editor-errors="ctrl.errors"
-                             data-file-editor-existing-items="ctrl.moderators"
-                             data-file-editor-item-validator="ctrl.constructor.addressValidator"
-                             data-file-editor-match-existing-items="true"
-                             data-file-editor-new-item="'mailing_list_common_add_email_addr' | translate"
+                        <div data-wuc-file-editor
+                             data-wuc-file-editor-errors="ctrl.errors"
+                             data-wuc-file-editor-existing-items="ctrl.moderators"
+                             data-wuc-file-editor-item-validator="ctrl.constructor.addressValidator"
+                             data-wuc-file-editor-match-existing-items="true"
+                             data-wuc-file-editor-new-item="'mailing_list_common_add_email_addr' | translate"
                              data-tr="tr"
                              data-trpl="trpl"
                              data-ng-model="ctrl.selection">
