+++ client/app/mailing-list/translations/Messages_fr_FR.xml
@@ -13,8 +13,6 @@
    <translation id="mailing_list_common_select_file" qtlid="261786">Sélectionner un fichier</translation>
    <translation id="mailing_list_common_select_unselect" qtlid="261799">Désélectionner</translation>
    <translation id="mailing_list_common_add_email_addr" qtlid="282575">Ajouter une adresse email</translation>
-   <translation id="mailing_list_common_addr_errors_1" qtlid="282588"><strong>1</strong> adresse en erreur</translation>
-   <translation id="mailing_list_common_addr_errors_other" qtlid="282601"><strong>{0}</strong> adresses en erreur</translation>
    <translation id="mailing_list_common_addr_delete_all" qtlid="159891">Tout supprimer</translation>
    <translation id="mailing_list_common_addr_go_to_next_err" qtlid="282562">Aller à l'erreur suivante</translation>
