+++ client/app/webUniverseComponents/fileEditor/file-editor.html
@@ -7,24 +7,24 @@
                 <span class="fa fa-upload" aria-hidden="true"></span>
             </label>
         </div>
-        <button type="button" class="btn btn-icon" title="{{:: 'common_add' | translate }}"
+        <button type="button" class="btn btn-icon" title="{{:: 'file_editor_common_add' | translate }}"
                 data-ng-click="goToAddItem()"
                 data-ng-show="showButtonAdd">
             <span class="fa fa-plus" aria-hidden="true"></span>
         </button>
         <span data-ng-show="ngModel.length > 0">|</span>
-        <button type="button" class="btn btn-icon" title="{{:: 'common_delete_all' | translate }}"
+        <button type="button" class="btn btn-icon" title="{{:: 'file_editor_common_delete_all' | translate }}"
                 data-ng-click="removeAll()"
                 data-ng-show="ngModel.length > 0">
             <span class="fa fa-trash-o" aria-hidden="true"></span>
         </button>
-        <span data-ng-show="fileEditorErrors.length > 0">|</span>
-        <button type="button" class="btn btn-icon text-danger" title="{{:: 'common_go_to_next_error' | translate }}"
+        <span data-ng-show="wucFileEditorErrors.length > 0">|</span>
+        <button type="button" class="btn btn-icon text-danger" title="{{:: 'file_editor_common_go_to_next_error' | translate }}"
                 data-ng-click="goToNextError()"
-                data-ng-show="fileEditorErrors.length > 0">
+                data-ng-show="wucFileEditorErrors.length > 0">
             <span class="fa fa-exclamation-triangle mr-2" aria-hidden="true"></span>
-            <span data-ng-bind-html="(fileEditorErrors.length > 1 ? 'mailing_list_common_addr_errors_other' : 'mailing_list_common_addr_errors_1') | translate: { t0: fileEditorErrors.length }"
-                  data-ng-show="fileEditorErrors.length > 0">
+            <span data-ng-bind-html="(wucFileEditorErrors.length > 1 ? 'file_editor_mailing_list_common_addr_errors_other' : 'file_editor_mailing_list_common_addr_errors_1') | translate: { t0: fileEditorErrors.length }"
+                  data-ng-show="wucFileEditorErrors.length > 0">
             </span>
         </button>
     </div>
@@ -35,7 +35,7 @@
             <a href="#" id="add-new-item"
                editable-text="newItem.value"
                onshow="newItem.value = ''"
-               oncancel="newItem.value = fileEditorNewItem"
+               oncancel="newItem.value = wucFileEditorNewItem"
                onbeforesave="checkAddItem($data)"
                onaftersave="addItem($data)"
                data-ng-bind="newItem.value">
@@ -46,12 +46,12 @@
 
             <a href="#" data-editable-text="ngModel[$index]"
                data-ng-bind="item"
-               data-ng-show="fileEditorItemValidator(item)">
+               data-ng-show="wucFileEditorItemValidator(item)">
             </a>
 
             <a href="#" class="text-danger" id="err-{{$index}}" data-editable-text="ngModel[$index]"
                data-ng-bind="item"
-               data-ng-show="!fileEditorItemValidator(item)">
+               data-ng-show="!wucFileEditorItemValidator(item)">
             </a>
 
             <button type="button" class="btn btn-icon"
