+++ client/app/webUniverseComponents/fileEditor/file-editor.less
@@ -1,0 +1,33 @@
+[wuc-file-editor],
+[data-wuc-file-editor] {
+  .file-editor-toolbar {
+    background: #f6f7f5;
+    border: 0.2rem solid #e1e4df;
+    border-radius: 6px 6px 0 0;
+    height: 28px;
+
+    input[type="file"] {
+      opacity: 0;
+      position: absolute;
+      z-index: -1;
+    }
+
+    input[type="file"]:focus + label {
+      outline: 5px auto -webkit-focus-ring-color;
+      outline-offset: -2px;
+    }
+  }
+
+  .file-editor {
+    .editor {
+      background: #fff;
+      border-color: #e1e4df;
+      border-radius: 0 0 6px 6px;
+      border-style: solid;
+      border-width: 0 0.2rem 0.2rem;
+      height: 140px;
+      overflow: auto;
+      padding: 5px;
+    }
+  }
+}
