+++ package.json
@@ -38,7 +38,6 @@
     "@bower_components/angular-scroll-glue": "Luegg/angularjs-scroll-glue#^2.2.0",
     "@bower_components/angular-ui-router": "angular-ui/angular-ui-router-bower#^1.0.0",
     "@bower_components/angular-vs-repeat": "kamilkp/angular-vs-repeat#v1.1.7",
-    "@bower_components/angular-xeditable": "vitalets/angular-xeditable#^0.1.10",
     "@bower_components/bootstrap": "twbs/bootstrap#^3.3.7",
     "@bower_components/bs4": "twbs/bootstrap#v4.0.0-beta",
     "@bower_components/ckeditor": "ckeditor/ckeditor-releases#^4.7.2",
@@ -65,6 +64,7 @@
     "angular-translate-loader-partial": "^2.18.1",
     "angular-translate-loader-static-files": "^2.18.1",
     "angular-ui-bootstrap": "1.3.3",
+    "angular-xeditable": "^0.1.10",
     "animate.css": "^3.5.0",
     "at-internet-ui-router-plugin": "^1.0.0",
     "babel-core": "^6.9.1",
